CRUD post

+++ src/routes/posts/posts.controller.ts
@@ -3,19 +3,12 @@ import { PostsService } from './posts.service'
 import { Auth } from 'src/shared/decorators/auth.decorator'
 import { AuthOptionsType, AuthType } from 'src/shared/enums/auth-type.enum'
 import { ActiveUser } from 'src/shared/decorators/active-user.decorator'
-import { GetPostItemDto } from './posts.dto'
-// import { Auth } from 'src/shared/decorators/auth.decorator'
-// import { AuthType } from 'src/shared/enums/auth-type.enum'
+import { CreatePostBodyDto, GetPostItemDto, UpdatePostBodyDto } from './posts.dto'
 
 @Controller('posts')
 export class PostsController {
   constructor(private readonly postsService: PostsService) {}
 
-  @Post()
-  @Auth([AuthType.Bearer])
-  create(@Body() bodyCreate: any, @ActiveUser('userId') userId: number) {
-    return this.postsService.create(bodyCreate, userId)
-  }
   @Get()
   @Auth([AuthType.Bearer, AuthType.APIKey], { condition: AuthOptionsType.AND })
   findAll(@ActiveUser('userId') userId: number) {
@@ -24,18 +17,26 @@ export class PostsController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.postsService.findOne(id)
+    return this.postsService.findOne(id).then((res) => new GetPostItemDto(res || {}))
+  }
+
+  @Post()
+  @Auth([AuthType.Bearer])
+  create(@Body() bodyCreate: CreatePostBodyDto, @ActiveUser('userId') userId: number) {
+    return this.postsService.create(bodyCreate, userId).then((res) => new GetPostItemDto(res))
   }
 
   @Put(':id')
   @Auth(AuthType.Bearer)
-  update(@Param('id') id: string, @Body() bodyUpdate: any) {
-    return this.postsService.update(id, bodyUpdate)
+  update(@Param('id') id: string, @Body() bodyUpdate: UpdatePostBodyDto, @ActiveUser('userId') userId: number) {
+    return this.postsService
+      .update({ postId: Number(id), userId, body: bodyUpdate })
+      .then((res) => new GetPostItemDto(res))
   }
 
   @Delete(':id')
   @Auth(AuthType.Bearer)
-  remove(@Param('id') id: string) {
-    return this.postsService.remove(id)
+  remove(@Param('id') id: string, @ActiveUser('userId') userId: number) {
+    return this.postsService.remove({ postId: Number(id), userId })
   }
 }
